+++ examples/next-openai/app/api/use-chat-streamdata/route.ts
@@ -1,5 +1,5 @@
 import { openai } from '@ai-sdk/openai';
-import { StreamData, streamText } from 'ai';
+import { convertToCoreMessages, generateId, StreamData, streamText } from 'ai';
 
 // Allow streaming responses up to 30 seconds
 export const maxDuration = 30;
@@ -7,15 +7,23 @@ export const maxDuration = 30;
 export async function POST(req: Request) {
   const { messages } = await req.json();
 
-  // optional: use stream data
+  // use stream data
   const data = new StreamData();
   data.append('initialized call');
 
   const result = await streamText({
-    model: openai('gpt-4-turbo'),
-    messages,
+    model: openai('gpt-4o'),
+    messages: convertToCoreMessages(messages),
     onFinish() {
+      // message annotation:
+      data.appendMessageAnnotation({
+        id: generateId(), // e.g. id from saved DB record
+        other: 'information',
+      });
+
+      // call annotation:
       data.append('call completed');
+
       data.close();
     },
   });