+++ packages/ai/core/tool/tool.ts
@@ -37,8 +37,14 @@ Use descriptions to make the input understandable for the language model.
   /**
 An async function that is called with the arguments from the tool call and produces a result.
 If not provided, the tool will not be executed automatically.
+
+@args is the input of the tool call.
+@options.abortSignal is a signal that can be used to abort the tool call.
    */
-  execute?: (args: inferParameters<PARAMETERS>) => PromiseLike<RESULT>;
+  execute?: (
+    args: inferParameters<PARAMETERS>,
+    options: { abortSignal?: AbortSignal },
+  ) => PromiseLike<RESULT>;
 }
 
 /**
@@ -47,10 +53,16 @@ Helper function for inferring the execute args of a tool.
 // Note: special type inference is needed for the execute function args to make sure they are inferred correctly.
 export function tool<PARAMETERS extends Parameters, RESULT>(
   tool: CoreTool<PARAMETERS, RESULT> & {
-    execute: (args: inferParameters<PARAMETERS>) => PromiseLike<RESULT>;
+    execute: (
+      args: inferParameters<PARAMETERS>,
+      options: { abortSignal?: AbortSignal },
+    ) => PromiseLike<RESULT>;
   },
 ): CoreTool<PARAMETERS, RESULT> & {
-  execute: (args: inferParameters<PARAMETERS>) => PromiseLike<RESULT>;
+  execute: (
+    args: inferParameters<PARAMETERS>,
+    options: { abortSignal?: AbortSignal },
+  ) => PromiseLike<RESULT>;
 };
 export function tool<PARAMETERS extends Parameters, RESULT>(
   tool: CoreTool<PARAMETERS, RESULT> & {