+++ packages/mistral/src/mistral-chat-language-model.ts
@@ -186,9 +186,21 @@ export class MistralChatLanguageModel implements LanguageModelV1 {
 
     const { messages: rawPrompt, ...rawSettings } = args;
     const choice = response.choices[0];
+    let text = choice.message.content ?? undefined;
+
+    // when there is a trailing assistant message, mistral will send the
+    // content of that message again. we skip this repeated content to
+    // avoid duplication, e.g. in continuation mode.
+    const lastMessage = rawPrompt[rawPrompt.length - 1];
+    if (
+      lastMessage.role === 'assistant' &&
+      text?.startsWith(lastMessage.content)
+    ) {
+      text = text.slice(lastMessage.content.length);
+    }
 
     return {
-      text: choice.message.content ?? undefined,
+      text,
       toolCalls: choice.message.tool_calls?.map(toolCall => ({
         toolCallType: 'function',
         toolCallId: toolCall.id,
@@ -231,7 +243,8 @@ export class MistralChatLanguageModel implements LanguageModelV1 {
       promptTokens: Number.NaN,
       completionTokens: Number.NaN,
     };
-    let isFirstChunk = true;
+    let chunkNumber = 0;
+    let trimLeadingSpace = false;
 
     return {
       stream: response.pipeThrough(
@@ -245,11 +258,11 @@ export class MistralChatLanguageModel implements LanguageModelV1 {
               return;
             }
 
+            chunkNumber++;
+
             const value = chunk.value;
 
-            if (isFirstChunk) {
-              isFirstChunk = false;
-
+            if (chunkNumber === 1) {
               controller.enqueue({
                 type: 'response-metadata',
                 ...getResponseMetadata(value),
@@ -275,11 +288,36 @@ export class MistralChatLanguageModel implements LanguageModelV1 {
 
             const delta = choice.delta;
 
+            // when there is a trailing assistant message, mistral will send the
+            // content of that message again. we skip this repeated content to
+            // avoid duplication, e.g. in continuation mode.
+            if (chunkNumber <= 2) {
+              const lastMessage = rawPrompt[rawPrompt.length - 1];
+
+              if (
+                lastMessage.role === 'assistant' &&
+                delta.content === lastMessage.content.trimEnd()
+              ) {
+                // Mistral moves the trailing space from the prefix to the next chunk.
+                // We trim the leading space to avoid duplication.
+                if (delta.content.length < lastMessage.content.length) {
+                  trimLeadingSpace = true;
+                }
+
+                // skip the repeated content:
+                return;
+              }
+            }
+
             if (delta.content != null) {
               controller.enqueue({
                 type: 'text-delta',
-                textDelta: delta.content,
+                textDelta: trimLeadingSpace
+                  ? delta.content.trimStart()
+                  : delta.content,
               });
+
+              trimLeadingSpace = false;
             }
 
             if (delta.tool_calls != null) {