+++ packages/amazon-bedrock/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ai-sdk/amazon-bedrock
 
+## 0.0.30
+
+### Patch Changes
+
+- 59d1abf: feat (provider/bedrock): support Bedrock amazon.titan-embed-text-v1 and amazon.titan-embed-text-v2:0 embeddings
+
 ## 0.0.29
 
 ### Patch Changes