+++ packages/mistral/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/mistral
 
+## 0.0.44
+
+### Patch Changes
+
+- b13ecf7: chore (provider/mistral): update model ids
+
+## 0.0.43
+
+### Patch Changes
+
+- 843656c: feat (provider/mistral): support continuation and trailing assistant messages
+
 ## 0.0.42
 
 ### Patch Changes