+++ packages/openai/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/openai
 
+## 0.0.68
+
+### Patch Changes
+
+- 741ca51: feat (provider/openai): support mp3 and wav audio inputs
+
+## 0.0.67
+
+### Patch Changes
+
+- 39fccee: feat (provider/openai): provider name can be changed for 3rd party openai compatible providers
+
 ## 0.0.66
 
 ### Patch Changes