+++ packages/react/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @ai-sdk/react
 
+## 0.0.64
+
+### Patch Changes
+
+- 98a3b08: fix (ui/react): fix callback dependency
+
+## 0.0.63
+
+### Patch Changes
+
+- caedcda: feat (ai/ui): add setData helper to useChat
+
 ## 0.0.62
 
 ### Patch Changes